Phase 1 consolidation: AI Hub, System Admin Hub, and merged guide configs
Major Changes:
- Created AI Hub (/ai-hub) consolidating 4 AI pages into tabs:
  * AI Assistant (chat interface & codebase indexing)
  * AI Enhanced Devices (device management)
  * AI Configuration (provider testing & setup)
  * API Keys (key management)

- Created System Admin Hub (/system-admin) consolidating 4 system pages into tabs:
  * Logs (log viewing & analytics)
  * Backup/Restore (backup management)
  * Config Sync (GitHub sync)
  * Tests (Wolf Pack matrix tests)

- Merged Guide Configs:
  * Sports Guide now has Configuration tab (links to sports-guide-config for full panel)
  * TV Guide now has Configuration tab with embedded config panel

- Updated main dashboard to reflect new structure:
  * AI Assistant → AI Hub
  * System Logs → System Admin
  * Removed redundant config links

- Removed old pages:
  * ai-assistant, ai-enhanced-devices, ai-config, ai-keys
  * logs, backup-restore, config-sync, tests
  * sports-guide-config, tv-guide-config (kept as standalone for deep config)

- Maintained all existing functionality
- Consistent color scheme and styling across all pages
- Improved navigation and user experience
